// ui/src/core/shared/types/core/country.ts
export type Country =
  | 'united-kingdom'
  | 'ukraine'
  | 'france'
  | 'argentina'
  | 'australia'
  | 'austria'
  | 'belgium'
  | 'brazil'
  | 'canada'
  | 'finland'
  | 'hungary'
  | 'ireland'
  | 'italy'
  | 'latvia'
  | 'poland'
  | 'scotland'
  | 'sweden'
  | 'taiwan'
  | 'united-arab-emirates'
  | 'germany'
  | 'united-states-of-america';